requirements.txt
standard-imghdr
python-telegram-bot==13.*
requests==2.32.*
python-dotenv==1.2.*
urllib3<2